requirements.txt
Flask==2.2.3
mysql-connector-python==8.0.33
python-dotenv==1.0.0
Flask-WTF==1.2.1
Flask-Login==0.6.2
requests==2.28.2
Werkzeug==2.2.3
APScheduler==3.10.1
homeassistant-api==4.2.2
